Scope processing algorithm: Return a default URL instead of undefined.
The default URL is the start URL with its filename, query and fragment
removed. This is consistent with Chrome's implementation and several
sites rely on this. It seems to be the most logical behaviour.

This means it is not possible to get an "unbounded" scope. Any reference
to unbounded scope in the document is now obsolete.