scope algorithm: Delete same-origin-as-document check.
This was redundant, because a) the start_url is already checked to be
same-origin-as-document, and b) the start_url is checked to be within
scope of scope (which implies they are same-origin). Thus the scope is
guaranteed to be same-origin as the document without this explicit
check.

This removes the document URL parameter to the scope algorithm.

No normative changes.