Breaking: Ignore any icon with no recognised purpose.
Previously, if all of an icon's purposes were unrecognised, it could be
used for any purpose. This causes forwards-compatibility issues, since
if we add a new purpose (such as the recently added 'maskable'), old
user agents will interpret icons with that new purpose as 'any', which
is not intended.

Icons with no stated purpose, or whose purpose is the empty string, are
still treated as 'any'.

Closes #728.