Processing a manifest: Moved defaults into individual processing steps.
This keeps the relevant info about the default values of members closely
related to the other relevant material about that member. It's also
consistent with how the rest of the members (e.g. scope) treat default
values. And prevents possible errors where the incorrect default value
is used by an intermediate step in between assigning the default and
assigning the actual value.