Rewrote Proprietary manifest members section to discourage vendor prefixing.
Previously, this section explicitly asked developers to add vendor
prefixing, which is unhealthy for the web long-term, and I don't believe
was actually the intention of this section. There is a subtle difference
between a feature linked to a proprietary ecosystem and a vendor prefix.
The newly written text makes this distinction clear, and allows
proprietary ecosystem features whilst discouraging vendor prefixing.
Updated examples to match.