Editorial: Replace ref to obsoleted RFC7540 with ref to RFC9110 (#1106)
The spec mentions RFC7540 when it talks about content negotiation. RFC7540 has
been obsoleted by RFC9113, so reference should be updated.

Now, the prose is a bit clunky because it seems to suggest that RFC7540 defines
the `"Accept-Language"` HTTP header, whereas it does not. Rather than replacing
the reference with a reference to RFC9113, this update rather targets RFC9110,
where `"Accept-Language"` is defined.